# api.yml
openapi: 3.0.1
info:
  title: ExtranetAPI
  version: '1.0'
paths:
...
  /projects/{projectId}:
    get:
      tags:
        - Project
      summary: Получить проект по идентификатору
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Получить проект по идентификатору
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProjectDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
    delete:
      tags:
        - Project
      summary: Удалить проект
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Удалить проект
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
  /projects:
    get:
      tags:
        - Project
      summary: Получить все проекты
      responses:
        '200':
          description: Получить все проекты
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDto'
    post:
      tags:
        - Project
      summary: Создать проект
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectDto'
          text/json:
            schema:
              $ref: '#/components/schemas/ProjectDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ProjectDto'
        required: true
      responses:
        '200':
          description: Создать проект
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
  /projects/{projectId}/project-updating:
    post:
      tags:
        - Project
      summary: Обновить основную информацию по проекту
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectDto'
          text/json:
            schema:
              $ref: '#/components/schemas/ProjectDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ProjectDto'
        required: true
      responses:
        '200':
          description: Обновить основную информацию по проекту
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
  /projects/{projectId}/contract-number:
    post:
      tags:
        - Project
      summary: Установить номер договора
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
        - name: contractNumber
          in: query
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Обновить номер договора
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
  /projects/{projectId}/start-date:
    post:
      tags:
        - Project
      summary: Установить актуальную дату начала выполнения
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
        - name: startDate
          in: query
          description: ''
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Установить актуальную дату начала выполнения
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
  /projects/{projectId}/end-date:
    post:
      tags:
        - Project
      summary: Установить актуальную дату завершения выполнения
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
        - name: endDate
          in: query
          description: ''
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Установить актуальную дату завершения выполнения
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
  /projects/{projectId}/deadline:
    post:
      tags:
        - Project
      summary: Установить ожидаемую дату завершения выполнения
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
        - name: deadLine
          in: query
          description: ''
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Установить ожидаемую дату завершения выполнения
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
  /projects/{projectId}/complete:
    post:
      tags:
        - Project
      summary: Завершить проект
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Завершить проект
  /projects/{projectId}/stop:
    post:
      tags:
        - Project
      summary: Остановить проект
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Остановить проект
  /projects/{projectId}/run:
    post:
      tags:
        - Project
      summary: Возобновить проект
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Возобновить проект
  /projects/{projectId}/summary-table:
    post:
      tags:
        - Project
      summary: Получить совдую таблицу по проекту
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Получить сводную таблицу
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProjectSummaryTableDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSummaryTableDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ProjectSummaryTableDto'
...
components:
  schemas:
...
    CostDto:
      type: object
      properties:
        name:
          type: string
          nullable: true
        amount:
          type: number
          format: double
        persent:
          type: number
          format: double
      additionalProperties: false
...
    CostPriceDto:
      type: object
      properties:
        costPrice:
          type: number
          format: double
        costs:
          type: array
          items:
            $ref: '#/components/schemas/CostDto'
          nullable: true
      additionalProperties: false
...
    ProjectDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        projectType:
          type: string
          nullable: true
        contractNumber:
          type: string
          nullable: true
        dateOfStart:
          type: string
          format: date-time
          nullable: true
        dateOfApplication:
          type: string
          format: date-time
        deadLine:
          type: string
          format: date-time
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        clientId:
          type: integer
          format: int32
        address:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        isCompleted:
          type: boolean
      additionalProperties: false
...
    ProjectSummaryTableDto:
      type: object
      properties:
        contractNumber:
          type: string
          nullable: true
        projectType:
          type: string
          nullable: true
        startDate:
          type: string
          format: date-time
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        numberOfDays:
          type: integer
          format: int32
        projectCost:
          type: number
          format: double
        costPrice:
          $ref: '#/components/schemas/CostPriceDto'
        margin:
          type: number
          format: double
        profitNorm:
          type: number
          format: double
        rateOfSurplusValue:
          type: number
          format: double
        stages:
          type: array
          items:
            $ref: '#/components/schemas/StageDto'
          nullable: true
      additionalProperties: false
    StageDto:
      type: object
      properties:
        name:
          type: string
          nullable: true
        isCompleted:
          type: boolean
      additionalProperties: false
...